requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg[binary]==3.2.13
psycopg-pool==3.2.8
python-dotenv==1.0.1
PyJWT==2.9.0
pytest==8.3.3
httpx==0.27.2
